Modify api-gw kubectl delete handler to always hash object name
The api-gw kubectl delete handler was only hashing the name of the object
to be deleted if the client is not-kubectl. This causes deletes through
kubectl to not work if the delete is being requested through object display
name.

This change is to always hash object name, that is being deleted.
This make the behavior uniform for all client of the api-gw.