Clear proxy env variabls for internal communication among containers
These changes clear proxy related env variables on containers that talk
internally.

The proxy env settings depends on the env in which the components are executing.
Regardless of env settings, we do not need the proxy env variables being set
as the containers involved in the runtime only communicate locally.
So this change resets the proxy env on all such containers.