Clear proxy env variabls for api-gw container
These changes clear proxy related env variables on api-gw container.

The proxy env settings depends on the env in which the components are executing.
Regardless of env settings, we do not need the proxy env variables being set
on the api-gw as it only communicates locally.
So this change resets the proxy env on the api-gw container.